Add files via upload
I completed the front end and part of the back end (One step left) for Flights, Hotels, and Car Rentals pages. Click the tabs in the navigation bar and it will take you to the correct pages. I added input boxes that align correctly with the API parameters specific to each section. I also added a button to each page so that when they are clicked, the API request is made.