Add files via upload
These are the files needed for the login and signup system. also I did change the tripmate home page name to just home in the files you can change it. however you would have to change it in every other file as well. also you need to a server using django for this to function.